android: wpa_s template: set p2p_go_max_inactivity and p2p concurrency
Set p2p_go_max_inactivity to 60 secs (instead of the default
of 300 secs) to avoid creating a user of experience of a
WiFi Direct connection which doesn't disconnect. This may
happen if the P2P GO interface misses a deauth/disassoc frame
and keeps on thinking the client is connected. Limit this
time to 30 secs which creates a better user experience.

[Arik - also set the default mode of operation to p2p concurrency. This
means we will work with 2 interfaces for p2p]

Signed-off-by: Eyal Shapira <eyal@wizery.com>
Signed-off-by: Arik Nemtsov <arik@wizery.com>