P2P: Fix broken rx action frame handling (INTERNAL-SQUASH)
The root cause of the duplication lies within the Rx path in mac80211
in Johannes` patch "mac80211: accept public action frames with mismatched BSSID".
When wpa_s works with a separate group interface this creates a duplicated Rx
on both the P2P dev interface and the GO interface.

Fix the issue by canceling the duplication in wpa_s and discarding
public action frames with unmatching bssid arriving on a dedicated GO interface

[Arik - for upstream we can't do this since it can break non-mac80211
drivers. Maybe set a flag in wiphy capabilities for when the kernel driver
does this?]

Signed-off-by: Eyal Shapira <eyal@wizery.com>